download editor files

+++ IndustrialPark/MainForm/MainForm.cs
@@ -8,6 +8,10 @@ using System.Linq;
 using System.Windows.Forms;
 using Newtonsoft.Json;
 using System.IO.Compression;
+using System.Web;
+using System.Net;
+using System.Threading.Tasks;
+using System.ComponentModel;
 
 namespace IndustrialPark
 {
@@ -47,7 +51,7 @@ namespace IndustrialPark
                 autoLoadOnStartupToolStripMenuItem.Checked = settings.AutoloadOnStartup;
                 checkForUpdatesOnStartupToolStripMenuItem.Checked = settings.CheckForUpdatesOnStartup;
 
-                if (settings.CheckForUpdatesOnStartup && UpdateIndustrialPark())
+                if (settings.CheckForUpdatesOnStartup && UpdateIndustrialPark(out _))
                 {
                     Close();
                     System.Diagnostics.Process.Start(Application.StartupPath + "\\IndustrialPark.exe");
@@ -214,72 +218,85 @@ namespace IndustrialPark
 
         private void CheckForUpdatesNowToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            if (UpdateIndustrialPark())
+            if (UpdateIndustrialPark(out bool hasChecked))
             {
                 Close();
                 System.Diagnostics.Process.Start(Application.StartupPath + "\\IndustrialPark.exe");
             }
+            else if (hasChecked)
+                MessageBox.Show("No update found.");
         }
 
-        public bool UpdateIndustrialPark()
+        public bool UpdateIndustrialPark(out bool hasChecked)
         {
-            string versionInfoURL = "https://raw.githubusercontent.com/igorseabra4/IndustrialPark/master/IndustrialPark/Resources/ip_version.json";
+            hasChecked = false;
 
             try
             {
-                using (var webClient = new System.Net.WebClient())
+                string versionInfoURL = "https://raw.githubusercontent.com/igorseabra4/IndustrialPark/master/IndustrialPark/Resources/ip_version.json";
+
+                string updatedJson;
+
+                HttpWebRequest request = (HttpWebRequest)WebRequest.Create(versionInfoURL);
+                request.AutomaticDecompression = DecompressionMethods.GZip;
+                using (HttpWebResponse response = (HttpWebResponse)request.GetResponse())
+                using (StreamReader reader = new StreamReader(response.GetResponseStream()))
+                    updatedJson = reader.ReadToEnd();
+                
+                IPversion updatedVersion = JsonConvert.DeserializeObject<IPversion>(updatedJson);
+                IPversion oldVersion = new IPversion();
+
+                hasChecked = true;
+
+                if (oldVersion.version != updatedVersion.version)
                 {
-                    string updatedJson = webClient.DownloadString(versionInfoURL);
+                    string messageText = "There is an update available for Industrial Park: " + updatedVersion.versionName + ". Do you wish to download it?";
+                    DialogResult d = MessageBox.Show(messageText, "Update Available", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
 
-                    IPversion updatedVersion = JsonConvert.DeserializeObject<IPversion>(updatedJson);
-                    IPversion oldVersion = new IPversion();
-
-                    if (oldVersion.version != updatedVersion.version)
+                    if (d == DialogResult.Yes)
                     {
-                        string messageText = "There is an update available for Industrial Park: " + updatedVersion.versionName + ". Do you wish to download it?";
-                        DialogResult d = MessageBox.Show(messageText, "Update Available", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
-
                         string updatedIPfileName = "IndustrialPark_" + updatedVersion.version.Replace('p', 'P') + ".zip";
                         string updatedIPURL = "https://github.com/igorseabra4/IndustrialPark/releases/download/" + updatedVersion.version + "/" + updatedIPfileName;
 
-                        if (d == DialogResult.Yes)
+                        string updatedIPfilePath = Application.StartupPath + "\\Resources\\" + updatedIPfileName;
+
+                        using (var webClient = new WebClient())
+                            webClient.DownloadFile(updatedIPURL, updatedIPfilePath);
+
+                        string oldPath = Application.StartupPath + "\\IndustrialPark_old\\";
+
+                        if (!Directory.Exists(oldPath))
+                            Directory.CreateDirectory(oldPath);
+
+                        foreach (string s in new string[]
                         {
-                            string oldPath = Application.StartupPath + "\\IndustrialPark_old\\";
-
-                            if (!Directory.Exists(oldPath))
-                                Directory.CreateDirectory(oldPath);
-
-                            foreach (string s in new string[]
-                            {
                                 "",
                                 "\\Resources",
                                 "\\Resources\\importvcolorobj",
                                 "\\Resources\\Models",
                                 "\\Resources\\SharpDX"
-                            })
+                        })
+                        {
+                            if (!Directory.Exists(oldPath + s))
+                                Directory.CreateDirectory(oldPath + s);
+
+                            foreach (string s2 in Directory.GetFiles(Application.StartupPath + s))
                             {
-                                if (!Directory.Exists(oldPath + s))
-                                    Directory.CreateDirectory(oldPath + s);
+                                if (Path.GetExtension(s2).ToLower().Equals(".zip"))
+                                    continue;
 
-                                foreach (string s2 in Directory.GetFiles(Application.StartupPath + s))
-                                {
-                                    string newFilePath = oldPath + s + "\\" + Path.GetFileName(s2);
+                                string newFilePath = oldPath + s + "\\" + Path.GetFileName(s2);
 
-                                    if (File.Exists(newFilePath))
-                                        File.Delete(newFilePath);
+                                if (File.Exists(newFilePath))
+                                    File.Delete(newFilePath);
 
-                                    File.Move(s2, newFilePath);
-                                }
+                                File.Move(s2, newFilePath);
                             }
-
-                            string updatedIPfilePath = Application.StartupPath + "\\Resources\\" + updatedIPfileName;
-                            webClient.DownloadFile(updatedIPURL, updatedIPfilePath);
-
-                            ZipFile.ExtractToDirectory(updatedIPfilePath, Application.StartupPath);
-                            
-                            return true;
                         }
+                        
+                        ZipFile.ExtractToDirectory(updatedIPfilePath, Application.StartupPath);
 
+                        return true;
                     }
                 }
             }
@@ -289,6 +306,44 @@ namespace IndustrialPark
             }
 
             return false;
+        }
+
+        private void DownloadIndustrialParkEditorFilesToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                string editorFilesZipURL = "https://github.com/igorseabra4/IndustrialPark-EditorFiles/archive/master.zip";
+                string destZipPath = Application.StartupPath + "\\Resources\\IndustrialPark-EditorFiles.zip";
+                string editorFilesFolder = Application.StartupPath + "\\Resources\\IndustrialPark-EditorFiles\\";
+                
+                using (var webClient = new WebClient())
+                    webClient.DownloadFile(new Uri(editorFilesZipURL), destZipPath);
+
+                RecursiveDelete(editorFilesFolder);
+
+                ZipFile.ExtractToDirectory(destZipPath, editorFilesFolder);
+
+                File.Delete(destZipPath);
+
+                MessageBox.Show("Downloaded newest IndustrialPark-EditorFiles to " + editorFilesFolder);
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }
+        private static void RecursiveDelete(string directory)
+        {
+            if (!Directory.Exists(directory))
+                return;
+
+            foreach (var dir in Directory.GetDirectories(directory))
+                RecursiveDelete(dir);
+
+            foreach (var s in Directory.GetFiles(directory))
+                File.Delete(s);
+
+            Directory.Delete(directory);
         }
 
         public ProjectJson FromCurrentInstance()
